% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-quantWarming.R
\docType{package}
\name{quantWarming}
\alias{quantWarming}
\title{quantWarming: Data and functions for trend analysis of temperature time series}
\description{
The quantWarming package provides datasets and functions for
trend analysis of temperature time series to reproduce the results
reported in Haupt and Fritsch (2021). Details on the included
datasets are provided in Haupt and Fritsch (2021).
}
